\name{bq_auth_for_rmd}
\alias{bq_auth_for_rmd}
\title{Big Query Authentication for running inside RMarkdown code chunks}
\usage{
bq_auth_for_rmd()
}
\description{
Supports reusing authentication tokens via an environment variable setting, so user doesn't have to authentiate with Google every time
}
\examples{
bq_auth_for_rmd()
}
